Mudança na estrutura de testes.
