--- api/apps/pessoas/tests/fixtures/clients.py
@@ -1,6 +1,0 @@
-import pytest
-
-@pytest.fixture
-def nao_autenticado_client():
-    from rest_framework.test import APIClient
-    return APIClient()
