+++ api/conftest.py
@@ -1,5 +1,4 @@
 pytest_plugins = [
-    'pessoas.tests.fixtures.clients',
-    'pessoas.tests.fixtures.urls',
-    'pessoas.tests.fixtures.pessoas_fixtures',
+    'tests.fixtures.utilidades',
+    'tests.fixtures.pessoas_fixtures',
 ]
