+++ api/tests/mocks/cliente_mock.py
