+++ api/tests/mocks/funcionario_mock.py
